"Renamed lipics template (currently not supported).
This renaming should clearly communicate that this setup is not a (well)
supported configuration. Firstly, lipics requires a recent version of
comment.sty that conflicts with the outdated version required by Isabelle
and, moreover, the lipics classes are non-free and, hence, prevent us
from distribution a patched version that works with Isabelle.